Authorization and Authentication Done
However there is a bug where I can't add a new item.

+++ project.py
@@ -34,6 +34,137 @@ def catalogJSON():
     return jsonify(Categories=[c.serialize for c in categories])
 
 
+@app.route('/catalog/login')
+def showLogin():
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in range(32))
+    login_session['state'] = state
+    return render_template('login.html', state=login_session['state'])
+
+
+@app.route('/catalog/connect', methods=['POST'])
+def connect():
+    if request.args.get('provider') == 'gplus':
+        return googleConnect()
+    flash("bad provider")
+    return redirect('/')
+
+
+def googleConnect():
+      # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code
+    code = request.data
+
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1])
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print "Token's client ID does not match app's."
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_credentials = login_session.get('credentials')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_credentials is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already connected.'),
+                                 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['credentials'] = credentials
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['email'] = data['email']
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 150px; height: 150px;border-radius: 75px;-webkit-border-radius: 75px;-moz-border-radius: 75px;"> '
+    flash("you are now logged in as %s" % login_session['username'])
+    print "done!"
+    return output
+
+
+@app.route('/catalog/disconnect')
+def disconnect():
+    # Only disconnect a connected user.
+    credentials = login_session.get('credentials')
+    if credentials is None:
+        response = make_response(
+            json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = credentials.access_token
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+
+    if result['status'] == '200':
+        # Reset the user's sesson.
+        del login_session['credentials']
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return redirect("/")
+    else:
+        # For whatever reason, the given token was invalid.
+        response = make_response(
+            json.dumps('Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        flash('You are now logged out.')
+        return redirect('/')
+
+
 @app.route('/')
 @app.route('/catalog')
 def showCatalog():
@@ -68,21 +199,26 @@ def showItem(cat_name, item_name):
     item = getItem(item_name)
     if not item:
         return redirect('NotFound')
-        
-    return render_template('item.html', item=item)
+
+    creator = getUserInfo(item.user_id)
+    return render_template('item.html', item=item, creator=creator)
 
 
 @app.route('/catalog/<cat_name>/item/add', methods=['GET', 'POST'])
 def addItem(cat_name):
+    if 'username' not in login_session:
+        return redirect('/catalog/login')
     category = getCategory(cat_name)
     categories = getCategories()
     if not category:
         return redirect('NotFound')
     if request.method == 'POST':
         try:
-            print request
+            user_id = getUserId(login_session['email'])
+            print user_id
             print request.form
-            newItem = Item(user_id=1,
+            print login_session
+            newItem = Item(user_id=user_id,
                            name=escape(request.form['item-name']),
                            price=escape(request.form['item-price']),
                            thumbnail=escape(request.form['item-thumb']),
@@ -102,6 +238,8 @@ def addItem(cat_name):
 
 @app.route('/catalog/<cat_name>/<item_name>/edit', methods=['GET', 'POST'])
 def editItem(cat_name, item_name):
+    if 'username' not in login_session:
+        return redirect('/catalog/login')
     itemToEdit = getItem(item_name)
     if request.method == 'POST':
         if request.form['item-name']:
@@ -127,6 +265,8 @@ def editItem(cat_name, item_name):
 
 @app.route('/catalog/<cat_name>/<item_name>/delete', methods=['GET', 'POST'])
 def deleteItem(cat_name, item_name):
+    if 'username' not in login_session:
+        return redirect('catalog/login')
     itemToDelete = getItem(item_name)
     if request.method == 'POST':
         session.delete(itemToDelete)
@@ -189,6 +329,29 @@ def getItem(item_name):
         return item
     except:
         return []
+
+
+def createUser(login_session):
+    newUser = User(name=login_session['username'], 
+                   email=login_session['email'], 
+                   picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
 
 
 def serializeCatalog():
